+++ src/extension.ts
@@ -1,32 +1,47 @@
 'use strict';
-import { ExtensionContext, commands, window, workspace } from 'vscode';
+import { ExtensionContext, commands, window, workspace, WorkspaceConfiguration } from 'vscode';
 import { SymbolsTreeDataProvider } from './symbolsTreeDataProvider';
-import { View, Sort } from './globals/enums';
+import { View, Sort, configurationSettings } from './globals/enums';
+import TelemetryReporter from 'vscode-extension-telemetry';
+
+const extensionPrefix: string = 'vscode-symbolexplorer';
+const extensionVersion: string = '0.0.9';
+const key: string = '43646677-3947-4bf8-a4ce-f8f2750385c2';
+
+let reporter;
 
 export function activate(context: ExtensionContext) {
-    const extensionPrefix: string = 'vscode-symbolexplorer';
-    const enum configurationSettings {
-        showExplorer = 'showExplorer',
-        showExplorerDebug = 'showExplorerDebug'
-    }
+    reporter = new TelemetryReporter(extensionPrefix, extensionVersion, key);
+    context.subscriptions.push(reporter);
+    reporter.sendTelemetryEvent('extension activated!');
 
-    const showExplorer: boolean = workspace.getConfiguration(extensionPrefix).get(configurationSettings.showExplorer);
-    const showExplorerDebug: boolean = workspace.getConfiguration(extensionPrefix).get(configurationSettings.showExplorerDebug);
+    const config: WorkspaceConfiguration = workspace.getConfiguration(extensionPrefix);
+    const showExplorer: boolean = config.get(configurationSettings.showExplorer);
+    const showExplorerDebug: boolean = config.get(configurationSettings.showExplorerDebug);
+    const autoStart: boolean = config.get(configurationSettings.autoStart);
+    const autoStartDelay: number = config.get(configurationSettings.autoStartDelay);
 
-    setupActivityBarView(context);
+    reporter.sendTelemetryEvent('configs', {
+        'showExplorer': showExplorer,
+        'showExplorerDebug': showExplorerDebug,
+        'autoStart': autoStart,
+        'autoStartDelay': autoStartDelay
+    });
+
+    setupActivityBarView(context, autoStart, autoStartDelay);
 
     if (showExplorer) {
-        setupExplorerView(context);
+        setupExplorerView(context, autoStart, autoStartDelay);
     }
 
     if (showExplorerDebug) {
-        setupDebugView(context);
+        setupDebugView(context, autoStart, autoStartDelay);
     }
 }
 
-function setupDebugView(context: ExtensionContext) {
+function setupDebugView(context: ExtensionContext, autoStart: boolean, autoStartDelay: number) {
     // Setup Tree Data Provider
-    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.Debug);
+    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.Debug, autoStart, autoStartDelay, reporter);
     context.subscriptions.push(window.registerTreeDataProvider('symbolExplorerDebug', symbolsTreeDataProvider));
 
     // Setup Navigate Command
@@ -46,9 +61,9 @@ function setupDebugView(context: ExtensionContext) {
     context.subscriptions.push(sortDescSymbolCommand);
 }
 
-function setupExplorerView(context: ExtensionContext) {
+function setupExplorerView(context: ExtensionContext, autoStart: boolean, autoStartDelay: number) {
     // Setup Tree Data Provider
-    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.Explorer);
+    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.Explorer, autoStart, autoStartDelay, reporter);
     context.subscriptions.push(window.registerTreeDataProvider('symbolExplorer', symbolsTreeDataProvider));
 
     // Setup Navigate Command
@@ -68,9 +83,9 @@ function setupExplorerView(context: ExtensionContext) {
     context.subscriptions.push(sortDescSymbolCommand);
 }
 
-function setupActivityBarView(context: ExtensionContext) {
+function setupActivityBarView(context: ExtensionContext, autoStart: boolean, autoStartDelay: number) {
     // Setup Tree Data Provider
-    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.View);
+    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.View, autoStart, autoStartDelay, reporter);
     context.subscriptions.push(window.registerTreeDataProvider('symbolExplorerView', symbolsTreeDataProvider));
 
     // Setup Navigate Command
@@ -91,4 +106,6 @@ function setupActivityBarView(context: ExtensionContext) {
 }
 
 // this method is called when your extension is deactivated
-export function deactivate() { }
+export function deactivate() {
+    reporter.dispose();
+}